Extract test helpers into a shared scope to remove duplication.
