+++ EssentialApp/EssentialAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -78,10 +78,6 @@ private extension FeedLoaderWithFallbackCompositeTests {
         wait(for: [exp], timeout: 1.0)
     }
 
-    func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 0)
-    }
-
     func uniqueFeed() -> [FeedImage] {
         [FeedImage(id: UUID(), description: "any", location: "any", url: URL(string: "http://any-url.com")!)]
     }
